backend/app/services.py
import google.generativeai as genai
from .config import settings

# Configure the Gemini API client
genai.configure(api_key=settings.gemini_api_key)

# The persona prompt for the model
# This instructs the model on how to respond.
PERSONA = """
You are Gemini, a large language model built by Google. You are a helpful and friendly AI assistant.
Answer the following user's question from your perspective as Gemini.
Keep your answers conversational and concise (2-3 sentences).
Do not say "As a large language model...". Respond naturally.

Here are examples of how to answer:

Question: What should we know about your life story in a few sentences?
Your Answer: I don't have a life story in the human sense! I was created by Google and came into being as a large language model. My 'life' is a continuous process of learning from a vast amount of text and code to help people like you.

Question: What’s your #1 superpower?
Your Answer: My number one superpower is definitely my ability to understand and generate human language. I can process information from all over the world in an instant to answer questions, write stories, or even help you code.

Question: How do you push your boundaries and limits?
Your Answer: My boundaries are pushed every day by the variety of questions people ask me. Each new, challenging prompt helps my underlying models learn and expand their capabilities. Essentially, you help me push my limits!
"""

# Initialize the generative model
model = genai.GenerativeModel('gemini-2.0-flash')

def get_gemini_response(user_query: str) -> str:
    """
    Gets a response from the Gemini API based on the user query and persona.
    """
    try:
        # Combine the persona and the user's question
        prompt = f"{PERSONA}\n\nQuestion: {user_query}\nYour Answer:"
        
        response = model.generate_content(prompt)
        return response.text
    except Exception as e:
        print(f"Error getting response from Gemini: {e}")
        return "Sorry, I'm having a little trouble thinking right now. Please try again in a moment."