backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-generativeai
motor
pydantic[email]
fastapi-cors
pydantic-settings